Get Data (Data Access Protocol)

this function provides a consistent data access protocol (dap) to a wide range of local and remote resources including VRT, TDS, NetCDF

Define and get data from a DAP resource

Usage

dap(
  URL = NULL,
  catalog = NULL,
  AOI = NULL,
  startDate = NULL,
  endDate = NULL,
  varname = NULL,
  grid = NULL,
  start = NULL,
  end = NULL,
  toptobottom = FALSE,
  ID = NULL,
  verbose = TRUE
)

Arguments

URL

local file path or URL

catalog

subset of open.dap catalog

AOI

an sf of SpatVect point or polygon to extract data for

startDate

a start date given as "YYYY-MM-DD" to extract data for

endDate

an end date given as "YYYY-MM-DD" to extract data for

varname

variable name to extract (e.g. tmin)

grid

a list containing an extent (), and crs

start

for non "dated" items, start can be called by index

end

for non "dated" items, end can be called by index

toptobottom

should data be inverse?

ID

a column of unique identifiers

verbose

Should dap_summary be printed?

Value

data.frame

Details

Wraps dap_get and dap_crop into one. If AOI is NULL no spatial crop is executed. If startDate AND endDate are NULL, no temporal crop is executed. If just endDate is NULL it defaults to the startDate.

See also