Write netrc file

Write a netrc file that is valid for accessing urs.earthdata.nasa.gov

Usage

writeNetrc(
  login,
  password,
  machine = "urs.earthdata.nasa.gov",
  netrcFile = getNetrcPath(),
  overwrite = FALSE
)

Arguments

login

A character. Email address used for logging in on earthdata

password

A character. Password associated with the login.

machine

the machine you are logging into

netrcFile

A character. A path to where the netrc file should be written. By default will go to your home directory, which is advised

overwrite

A logical. overwrite the existing netrc file?

Value

A character vector containing the netrc file path

Details

The database is accessed with the user's credentials. A netrc file storing login and password information is required. See here. Once set up you must do the following (1) Login to EarthData (2) Go to Applications > Authorized Apps (3) If NASA GESDISC DATA ARCHIVE is not in the Approved Applications list, select APPROVE MORE APPLICATIONS (4) Find NASA GESDISC DATA ARCHIVE and click AUTHORIZE for instruction on how to register and set DataSpace credential.

See also

Examples

if (FALSE) { # \dontrun{
writeNetrc(
  login = "XXX@email.com",
  password = "yourSecretPassword"
)
} # }